Implementing database aside with API (both working correctly)

+++ pkg/id.go
@@ -8,6 +8,6 @@ func ParseIdToInt64(val int) ID {
 	return int64(val)
 }
 
-func ParseIdToString(val int) string {
-	return strconv.Itoa(val)
+func ParseIdToString(val int64) string {
+	return strconv.Itoa(int(val))
 }
